pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-junior-associate"
version = "1.0.0"
description = "A polished Python library for scraping legal case law from multiple jurisdictions"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "The Junior Associate Contributors", email = "contributors@thejuniorassociate.org"}]
keywords = ["legal", "case law", "scraping", "court", "judgments", "research"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.0",
    "urllib3>=2.0.0",
    "certifi>=2023.5.7",
    "charset-normalizer>=3.1.0",
    "idna>=3.4",
    "soupsieve>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[project.scripts]
junior-associate = "the_junior_associate.cli:main"

[project.urls]
Homepage = "https://github.com/gongahkia/the-junior-associate"
Repository = "https://github.com/gongahkia/the-junior-associate"
Documentation = "https://the-junior-associate.readthedocs.io/"
"Bug Reports" = "https://github.com/gongahkia/the-junior-associate/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=the_junior_associate --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]